Rewrite Plot and axis-like types. (#67)
* Rewrite Plot and axis-like types.

This is a step towards fixing #44 (almost there!).

1. New user-facing plot/plot3 constructor is

```julia
Plot([incrementa], [options], data, trailing...)
```
and similarly for `Plot3`.

`incremental` defaults to `false`, as this seems to be the most common
use case.

Some validation is done on `data` (checking for type).

This removes the `label` kwarg, and allows us to close #16. Examples
now recommend an explicit `\addlegendentry`.

Docstrings are added for everything, and examples are rewritten
accordingly.

2. Axis-like code cleaned up a bit with macros. Variations on log axes
added. Explicitly document that strings are emitted as is.

3. GroupPlot rewritten, allow multiple plots and empty \nextgroupplot,
two examples added, one from the manual and one for multiple plots.

4. Replaced random examples with deterministic ones, perhaps stick to
this and close #63? Minor indentation issues fixed.

5. Minor fixes for testing framework.

* Separate incremental plot constructors, add LegendEntry.

The `incremental` flag was removed from `Plot*` constructors in the
API, replaced by `PlotInc` and `Plot3Inc`.

A LegendEntry type was added.

Minor fixes:

- remove the INCREMENTAL constant, as it is now unnecessary

- clarify docstrings

- disabled docstring checks, as they lead to infinite loops for some
  reason

* Trivial docstring fix.